requirements.txt
et-xmlfile==1.1.0
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
click==8.1.3
